requirements.txt
anyio==3.3.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
blinker==1.4
certifi==2019.6.16
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.4
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.3
Flask-Cors==3.0.10
Flask-DebugToolbar==0.10.1
flask-paginate==0.8.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.8
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.2.3
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.3
MarkupSafe==1.1.1
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
optional-django==0.3.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pd==0.0.1
pickleshare==0.7.5
pipenv==2021.5.29
prometheus-client==0.9.0
prompt-toolkit==3.0.14
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.2
react==4.3.0
requests==2.22.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.20
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.4
zipp==3.4.0
